fix: consertar continuidade das migrações

+++ db/migrate/20230712004134_create_participante_provas.rb
@@ -1,7 +1,8 @@
 class CreateParticipanteProvas < ActiveRecord::Migration[7.0]
+  def up
+    drop_table :participante_provas if table_exists?(:participante_provas)
+  end
   def change
-    drop_table :participante_provas # Remove a tabela existente
-
     create_table :participante_provas, id: false do |t|
       t.references :participante, null: false, foreign_key: { to_table: :participantes }
       t.references :prova, null: false, foreign_key: { to_table: :provas }
