+++ db/migrate/20230712195643_remove_usuarios_id_and_rename_usuario_id_in_participantes.rb
@@ -1,6 +1,8 @@
 class RemoveUsuariosIdAndRenameUsuarioIdInParticipantes < ActiveRecord::Migration[7.0]
+  def up
+    remove_column :participantes, :usuarios_id if column_exists?(:participantes, :usuarios_id)
+  end
   def change
-    remove_column :participantes, :usuarios_id
     rename_column :participantes, :usuario_id, :usuarios_id
   end
 end
